Home tasks are refactored
