+++ Git/Git.csproj
@@ -7,4 +7,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="hm3\" />
+  </ItemGroup>
+
 </Project>
